improve example to read WiFi credentials from .txt
improve example to read WiFi credentials from a text file ('known_wifis.txt') that can be uploaded to SPIFFS memory, so that the user does not have to publish a code on his public GitHub repository that contains his private WiFi credentials. instead add the possibility to upload a text file to SPIFFS that contains the WiFi credentials, und the code will dynamically load the credentials from the .txt file that is stored locally on the device.