add WiFiManager and SSID presets in SPIFFS
Add the possibility to upload a text file named known_wifis.txt to SPIFFS memory to allow keeping sensitive WiFi credentials only locally on the device instead of hard coded in the sketch.
That prevents that Wifi credentials would be visible online in the Arduino code on a public GitHub repo, if someone uses this code.
If the file is not provided, or the ESP8266 cannot connect to any of the named WiFis in the file within a configurable time, start WifiManager to configure Wifi credentials. If that times out, proceed to void loop() after an adjustable timeout.
If a Wifi connection can be established,  start the OTA update of ESP8266 from the GitHub repository, using the latest release that has a correctly named binary uploaded as an asset of this latest release.