add feature: multiple reset starts Wifimanager
double reset detector (may need 3x reset) deletes wifi credentials stored in wifimanager (not the Wifis that are stored in preconfiguration file on SPIFFS memory), to allow starting wifimanager from new if necessary (e.g. if you are connected to a WiFi that you set, but want to use a different one).